Added a new password song, check desc for info
The song is a Peak Version of Trial

Basically trial but it's made in a similar way to the VS Harcvuk mod from 2021 - 2022 back when my art and music were quite shit. Friends and I always like to mock it bc it's funny, plus it shows just how much I've changed. but uh yeah