Too Peak now has a cutscene (READ DESC)
The code for it is kinda iffy af, and how so?

Well, that's bc I literally took it from Top 10 Inc just to make it work lmfao

You can go ahead and make a replacement that functions fine, this is meant as a placeholder of sorts anyway.
Just make sure it lets the cutscene play from the password menu and the song select menu. Basically let the cutscene play on every instance of playing the song (maybe except for chart editing)